Refactor witchlights-fastled to separate modules.
This should be a strict no-op refactor of existing code.  However it doesn't
preserve the order of items in memory.

There are still a lot of dangerous cross-dependencies and cross-module
access, but this provides a starting point for making the libraries more
testable.

TESTED=builds and links with ino.  Maybe it works?